refactor: Split out game logic to GameView extensions, prep for ace/score improvements
