+++ Blackjack-iOS/Shared/Enums/GamePhase.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+
+enum GamePhase {
+    case idle
+    case dealing
+    case playerTurn
+    case dealerTurn
+    case gameOver
+}
