+++ Blackjack-iOS/Shared/Enums/Outcome.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+
+enum Outcome {
+    case playerWins, dealerWins, push, playerBlackjack
+}
