--- Blackjack-iOS/Views/GameViewManager.swift
@@ -1,12 +1,0 @@
-// GameViewManager.swift
-// Blackjack-iOS
-//
-// Updated by ChatGPT on 2025-07-15.
-
-import Foundation
-import Observation
-
-@Observable
-class GameViewManager {
-   
-}
